fix - allow disabling logging error logs
